refactor: Move features to their own namespace (#439)
* Move features to their own namespace

* Fix up references that VS didn't do.
