+++ Childrens-Social-Care-CPD-Tests/Configuration/Features/FeaturesConfigBackgroundServiceTests.cs
@@ -1,10 +1,9 @@
 ﻿using Childrens_Social_Care_CPD.Configuration;
+using Childrens_Social_Care_CPD.Configuration.Features;
 using Microsoft.Extensions.Logging;
-using NSubstitute;
-using NUnit.Framework;
 using System.Threading;
 using System.Threading.Tasks;
-namespace Childrens_Social_Care_CPD_Tests.Configuration;
+namespace Childrens_Social_Care_CPD_Tests.Configuration.Features;
 
 public class FeaturesConfigBackgroundServiceTests
 {
