+++ Childrens-Social-Care-CPD/Configuration/Features/FeaturesConfig.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Concurrent;
 
-namespace Childrens_Social_Care_CPD.Configuration;
+namespace Childrens_Social_Care_CPD.Configuration.Features;
 
 public sealed class FeaturesConfig : IFeaturesConfig
 {
