+++ Childrens-Social-Care-CPD/Controllers/ResourcesController.cs
@@ -1,4 +1,5 @@
 ﻿using Childrens_Social_Care_CPD.Configuration;
+using Childrens_Social_Care_CPD.Configuration.Features;
 using Childrens_Social_Care_CPD.Contentful.Models;
 using Childrens_Social_Care_CPD.DataAccess;
 using Childrens_Social_Care_CPD.GraphQL.Queries;
