+++ Childrens-Social-Care-CPD/Views/Resources/Resource.cshtml
@@ -1,7 +1,7 @@
-﻿@using Childrens_Social_Care_CPD.Configuration
-@using Childrens_Social_Care_CPD.Contentful.Models;
+﻿@using Childrens_Social_Care_CPD.Contentful.Models;
 @using Childrens_Social_Care_CPD.Contentful.Renderers;
 @using Contentful.Core.Models;
+@using Childrens_Social_Care_CPD.Configuration.Features;
 
 @model Content;
