+++ Childrens-Social-Care-CPD/Views/SearchResources/SearchResources.cshtml
@@ -1,4 +1,4 @@
-﻿@using Childrens_Social_Care_CPD.Configuration
+﻿@using Childrens_Social_Care_CPD.Configuration.Features
 @using Childrens_Social_Care_CPD.Controllers
 @using Childrens_Social_Care_CPD.Models
 @using Microsoft.AspNetCore.Html
