+++ Childrens-Social-Care-CPD/Views/Shared/_Header.cshtml
@@ -1,4 +1,4 @@
-﻿@using Childrens_Social_Care_CPD.Configuration;
+﻿@using Childrens_Social_Care_CPD.Configuration.Features;
 
 @inject IFeaturesConfig featuresConfig
