+++ Childrens-Social-Care-CPD/WebApplicationBuilderExtensions.cs
@@ -20,6 +20,7 @@ using System.Net.Http.Headers;
 using Microsoft.AspNetCore.DataProtection;
 using Azure.Storage.Blobs;
 using Azure.Identity;
+using Childrens_Social_Care_CPD.Configuration.Features;
 
 namespace Childrens_Social_Care_CPD;
